SlateEditor: refactor into several component files
